update grid step size
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@2126 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/gl3d/camera/GL3DCameraOptionsPanel.java
@@ -81,8 +81,8 @@ public class GL3DCameraOptionsPanel extends JPanel implements GL3DCameraSelectio
 
     private void changeCamera(GL3DCamera newCamera) {
         boolean hidden = cameraSelectorModel.getCurrentCamera().getGrid().getDrawBits().get(Bit.Hidden);
-        int resx = cameraSelectorModel.getCurrentCamera().getGridResolutionX();
-        int resy = cameraSelectorModel.getCurrentCamera().getGridResolutionY();
+        double resx = cameraSelectorModel.getCurrentCamera().getGridResolutionX();
+        double resy = cameraSelectorModel.getCurrentCamera().getGridResolutionY();
 
         cameraSelectorModel.getCurrentCamera().deactivate();
         cameraSelectorModel.setCurrentCamera(newCamera);
